Added limits to html and body
After adding the Viewpoint entry into index.html, the mobile version did shrink down to the correct width of about 640px

but there was a problem where the mobile version would show a white space also of about 640px would appear directly to the right of the page. This should fix it.